requirements.txt
numpy>=1.26,<3
pandas>=2.2,<3
yfinance>=0.2.54,<0.3
ib_insync>=0.9,<1
tabulate>=0.9,<1
PyYAML>=6,<7
lxml>=5,<6
